+++ src/Endpoint/Contracts/Hello5/Models/Echo/HelloEchoConfigResponseDto.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-using CoreFX.Abstractions.Contracts;
-
-namespace Hello5.Domain.Contract.Models.Echo
-{
-    public class HelloEchoConfigResponseDto : SvcResponse<Dictionary<string, bool>>
-    {
-    }
-}