+++ src/Endpoint/Contracts/Hello5/Models/Echo/HelloEchoDBResponseDto.cs
@@ -1,8 +1,0 @@
-﻿using CoreFX.Abstractions.Contracts;
-
-namespace Hello5.Domain.Contract.Models.Echo
-{
-    public class HelloEchoDBResponseDto : SvcResponse<string>
-    {
-    }
-}