+++ src/Endpoint/Contracts/Hello5/Models/Echo/HelloEchoResponseDto.cs
@@ -1,7 +1,0 @@
-﻿namespace Hello5.Domain.Contract.Models.Echo
-{
-    public class HelloEchoResponseDto
-    {
-        public string Recv { get; set; }
-    }
-}