+++ src/Endpoint/Contracts/Hello8/Models/Echo/HelloEchoConfigResponseDto.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+using CoreFX.Abstractions.Contracts;
+
+namespace Hello8.Domain.Contract.Models.Echo
+{
+    public class HelloEchoConfigResponseDto : SvcResponse<Dictionary<string, bool>>
+    {
+    }
+}