+++ src/Endpoint/Contracts/Hello8/Models/Echo/HelloEchoDBResponseDto.cs
@@ -1,0 +1,8 @@
+﻿using CoreFX.Abstractions.Contracts;
+
+namespace Hello8.Domain.Contract.Models.Echo
+{
+    public class HelloEchoDBResponseDto : SvcResponse<string>
+    {
+    }
+}