+++ src/Endpoint/Contracts/Hello8/Models/Echo/HelloEchoRequestDto.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Hello8.Domain.Contract.Models.Echo
+{
+    public class HelloEchoRequestDto
+    {
+        [Required(AllowEmptyStrings = false)]
+        public string Send { get; set; }
+    }
+}