+++ src/Endpoint/Contracts/Hello8/Models/Echo/HelloEchoResponseDto.cs
@@ -1,0 +1,7 @@
+﻿namespace Hello8.Domain.Contract.Models.Echo
+{
+    public class HelloEchoResponseDto
+    {
+        public string Recv { get; set; }
+    }
+}