+++ src/Endpoint/Hello8/Controllers/Bases/HelloContollerBase.cs
@@ -1,0 +1,16 @@
+﻿using CoreFX.Abstractions.Logging;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+
+namespace Hello8.Domain.Endpoint.Controllers.Bases
+{
+    public abstract class HelloContollerBase : ControllerBase
+    {
+        protected readonly ILogger _logger;
+
+        protected HelloContollerBase()
+        {
+            _logger = LogMgr.CreateLogger(GetType());
+        }
+    }
+}