+++ ChangeLog.md
@@ -1,7 +1,11 @@
+### 2024-09-09
+* **Hello8.Domain.Endpoint (2.0.1)**
+  * refactor: Upgrade to .NET 8.0 from .NET 5.0
+
 ### 2020-11-30
 * **Hello5.Domain.Endpoint (1.0.0.2)**
   * feat: Support docker, k8s and runtime environment variables
-  
+
 * **Hello5.Domain.Contract (1.0.0.0)**
   * Created